org : removed bark-small example and refactored convert.py to accept multiple bark models
