+++ README.md
@@ -47,36 +47,32 @@ make -j && ./main -p "This is an audio generated by bark.cpp"
 /_.___/\__,_/_/  /_/|_|  (_)   \___/ .___/ .___/
                                   /_/   /_/
 
-
-bark_tokenize_input: prompt: 'this is a dog barking.'
-bark_tokenize_input: number of tokens in prompt = 513, first 8 tokens: 20579 20172 10217 27883 28169 25677 10167 129595
+bark_tokenize_input: prompt: 'This is an audio generated by bark.cpp'
+bark_tokenize_input: number of tokens in prompt = 513, first 8 tokens: 20795 20172 20199 33733 58966 20203 28169 20222
 
 Generating semantic tokens: [========>                                          ] (17%)
 
-bark_print_statistics: mem per token =     0.00 MB
-bark_print_statistics:   sample time =     9.90 ms / 138 tokens
-bark_print_statistics:  predict time =  3163.78 ms / 22.92 ms per token
-bark_print_statistics:    total time =  3188.37 ms
+bark_print_statistics:   sample time =    10.98 ms / 138 tokens
+bark_print_statistics:  predict time =   614.96 ms / 4.46 ms per token
+bark_print_statistics:    total time =   633.54 ms
 
 Generating coarse tokens: [==================================================>] (100%)
 
-bark_print_statistics: mem per token =     0.00 MB
-bark_print_statistics:   sample time =     3.96 ms / 410 tokens
-bark_print_statistics:  predict time = 14303.32 ms / 34.89 ms per token
-bark_print_statistics:    total time = 14315.52 ms
+bark_print_statistics:   sample time =     3.75 ms / 410 tokens
+bark_print_statistics:  predict time =  3263.17 ms / 7.96 ms per token
+bark_print_statistics:    total time =  3274.00 ms
 
 Generating fine tokens: [==================================================>] (100%)
 
-bark_print_statistics: mem per token =     0.00 MB
-bark_print_statistics:   sample time =    41.93 ms / 6144 tokens
-bark_print_statistics:  predict time = 15234.38 ms / 2.48 ms per token
-bark_print_statistics:    total time = 15282.15 ms
+bark_print_statistics:   sample time =    38.82 ms / 6144 tokens
+bark_print_statistics:  predict time =  4729.86 ms / 0.77 ms per token
+bark_print_statistics:    total time =  4772.92 ms
 
-Number of frames written = 51840.
+write_wav_on_disk: Number of frames written = 65600.
 
-main:     load time =  1436.36 ms
-main:     eval time = 34520.53 ms
-main:    total time = 32786.04 ms
+main:     load time =   324.14 ms
+main:     eval time =  8806.57 ms
+main:    total time =  9131.68 ms
 ```
 
 Here are typical audio pieces generated by `bark.cpp`:
